Added spinner. Refactored folder
