+++ src/App.jsx
@@ -1,10 +1,10 @@
 import React from "react";
 import { Routes, Route } from "react-router-dom";
-import ProtectedRoute from "./comps/ProtectedRoute";
+import ProtectedRoute from "./auth/ProtectedRoute";
 import Landing from "./pages/Landing";
 import Login from "./pages/Login";
 import Home from "./pages/Home";
-import { AuthProvider } from "./hooks/useAuth";
+import { AuthProvider } from "./auth/useAuth";
 
 const App = () => (
   <AuthProvider>
