+++ src/auth/ProtectedRoute.jsx
@@ -1,5 +1,5 @@
 import { Navigate } from "react-router-dom";
-import { useAuth } from "../hooks/useAuth";
+import { useAuth } from "./useAuth";
 
 export const ProtectedRoute = ({ children }) => {
   const user = useAuth();
